INFO works, GET almost works
INFO works perfect. GET works perfectly for files (writes contents of file to client). GET for directories works, but writes files that start with "." to client even though it shouldn't (don't know what to do about this). If pathname leads to invalid location (file or directory doesn't exist) then an error message is written to client.